feat: made primary action button dependent on context
When using the Workflow plugin, the primary action for the editor becomes 'Begin Workflow'.
This is a bit disruptive to the normal flow they're used to - starting a review workflow will
only happen in a minority of cases, making 'Publish' still the preferred primary action.
Also, after beginning a workflow, 'Assign' becomes the primary action regardless of if there are
already assignees or not.A better primary action in these circumstances is the next workflow state.
These changes attempt to create a better UX, by providing more relevant primary actions to the
user, depending on the current status of the document