feat: publish when completing workflow
It could be confusing for the editors that completing the workflow for a document
doesn't actually do anything to the document itself. It only removes the document
from the workflow.
Instead, completing the workflow for a document should also publish the document.
Also changed labels to convey the action more clearly to the user.